refactor: move protobuf test fixtures to dedicated testing file
