+++ tycho-indexer/src/lib.rs
@@ -3,6 +3,8 @@ pub mod extractor;
 pub mod pb;
 pub mod services;
 pub mod substreams;
+
+#[cfg(test)]
 mod testing;
 
 #[cfg(test)]
